chore: consolidates utils in app directory
