+++ app/_components/database-wrapper/index.tsx
@@ -12,7 +12,7 @@ import { useSearchParams } from "next/navigation"
 import xss from "xss"
 import { debounce } from "lodash-es"
 import { useFilter, useFetchGames } from "../../_utils/hooks"
-import { getNextUrlState } from "../../_utils/set-params"
+import { getNextUrlState } from "../../_utils/helpers"
 import { Search } from "./Search"
 import { FilterOptions } from "./FilterOptions"
 import { GameList } from "./GameList"
