+++ app/_components/route-tracker/index.tsx
@@ -2,7 +2,7 @@
 
 import { useEffect } from "react"
 import { usePathname } from "next/navigation"
-import { pageView } from "utils/page-view"
+import { pageView } from "@/app/_utils/helpers"
 
 export function RouteTracker() {
   const pathname = usePathname()
