+++ app/_utils/helpers.ts
@@ -1,5 +1,34 @@
 import { DatabaseGame, DatabasePlatform } from "types/games"
 import { isPlainObject, isString, pickBy } from "lodash-es"
+import { FilterItem } from "types/games"
+
+export function getNextUrlState({
+  searchValue,
+  filteredPlatforms,
+}: {
+  searchValue: string
+  filteredPlatforms: FilterItem[]
+}): URL {
+  const platformValues = filteredPlatforms
+    .filter((item) => item.selected)
+    .map((item) => item.value)
+
+  const url = new URL(window.location.origin)
+
+  if (searchValue) {
+    url.searchParams.set("search", searchValue)
+  } else if (url.searchParams.has("search")) {
+    url.searchParams.delete("search")
+  }
+
+  if (platformValues.length > 0) {
+    platformValues.forEach((p) => url.searchParams.append("platform", p))
+  } else if (url.searchParams.has("platform")) {
+    url.searchParams.delete("platform")
+  }
+
+  return url
+}
 
 export function sortByKey(key: string) {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -47,4 +76,12 @@ export function transformGameData(database: { platforms: DatabasePlatform[] }) {
   })
 
   return { games, queryData }
+}
+
+export function pageView(path_url: string) {
+  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+  // @ts-ignore
+  window?.gtag?.("config", process.env.NEXT_PUBLIC_MEASUREMENT_ID, {
+    path_url,
+  })
 }
