+++ app/constants.ts
@@ -1,3 +1,5 @@
+export const BASE_TITLE = "a few games |"
+
 export const DatabaseFields = {
   PRODUCT_CODE: "product_code",
   SATELLITE_CODE: "satellite_code",
@@ -32,4 +34,22 @@ export const DB_FIELD_DESCRIPTIONS: Record<string, string> = {
     "The manufacturers part number. Typically found on the back of a game near the UPC.",
   [DatabaseFields.NOTES]:
     "Extra discerning details such as packaging differences, errors, language, or the like.",
+}
+
+export const EMAIL_REGEXP: RegExp =
+  /^(([^<>()[\]\\.,;:\s@"]+(\.[^<>()[\]\\.,;:\s@"]+)*)|(".+"))@(([[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/
+
+export const ERROR_MESSAGE: string =
+  "There was an error subscribing to the newsletter. Try again or email contact@afew.games if the error persists."
+
+export const SubscribeFormStatuses: {
+  ERROR: "error"
+  SUCCESS: "success"
+  LOADING: "loading"
+  NONE: "none"
+} = {
+  ERROR: "error",
+  SUCCESS: "success",
+  LOADING: "loading",
+  NONE: "none",
 }
