+++ package.json
@@ -42,6 +42,7 @@
     "react-dom": "^18.2.0",
     "react-google-recaptcha-v3": "^1.10.1",
     "react-tooltip": "^5.25.1",
+    "server-only": "^0.0.1",
     "tinacms": "^2.2.8",
     "xss": "^1.0.14"
   },
